python/evm-message-signing/pyproject.toml
[project]
name = "python-signer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "aiohttp>=3.10.11",
    "ecdsa>=0.19.0",
    "eth-account>=0.13.5",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
